LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/parsing/scroll-padding-block-inline-shorthand-expected.txt


FAIL e.style['scroll-padding-block'] = "auto 10px" should set scroll-padding-block-end assert_equals: scroll-padding-block-end should be canonical expected (string) "10px" but got (undefined) undefined
FAIL e.style['scroll-padding-block'] = "auto 10px" should set scroll-padding-block-start assert_equals: scroll-padding-block-start should be canonical expected (string) "auto" but got (undefined) undefined
FAIL e.style['scroll-padding-block'] = "auto 10px" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['scroll-padding-block'] = "20%" should set scroll-padding-block-end assert_equals: scroll-padding-block-end should be canonical expected (string) "20%" but got (undefined) undefined
FAIL e.style['scroll-padding-block'] = "20%" should set scroll-padding-block-start assert_equals: scroll-padding-block-start should be canonical expected (string) "20%" but got (undefined) undefined
FAIL e.style['scroll-padding-block'] = "20%" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['scroll-padding-inline'] = "10px auto" should set scroll-padding-inline-end assert_equals: scroll-padding-inline-end should be canonical expected (string) "auto" but got (undefined) undefined
FAIL e.style['scroll-padding-inline'] = "10px auto" should set scroll-padding-inline-start assert_equals: scroll-padding-inline-start should be canonical expected (string) "10px" but got (undefined) undefined
FAIL e.style['scroll-padding-inline'] = "10px auto" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['scroll-padding-inline'] = "0%" should set scroll-padding-inline-end assert_equals: scroll-padding-inline-end should be canonical expected (string) "0%" but got (undefined) undefined
FAIL e.style['scroll-padding-inline'] = "0%" should set scroll-padding-inline-start assert_equals: scroll-padding-inline-start should be canonical expected (string) "0%" but got (undefined) undefined
FAIL e.style['scroll-padding-inline'] = "0%" should not set unrelated longhands assert_true: expected true got false
